refactor constant
