+++ lib/SearchPage/SearchPage.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:grocery/SearchPage/SearchBar.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 
 // ignore: must_be_immutable
 class SearchPage extends StatelessWidget {
