+++ lib/constants/Application.dart
@@ -1,0 +1,6 @@
+import 'package:grocery/constants/ApplicationColor.dart';
+
+class Application {
+  static final double defaultPadding = 20.0;
+  static final String httBaseUrl = '';
+}
