+++ lib/constants/ApplicationColor.dart
@@ -1,0 +1,11 @@
+import 'dart:ui';
+
+class ApplicationColor{
+  static final primaryColor = Color.fromRGBO(255, 132, 76, 1);
+  static final textColor = Color.fromRGBO(0, 0, 0, 1);
+  static final backGroundColor = Color.fromRGBO(237, 243, 242, 1);
+  static final naturalWhite = Color.fromRGBO(245, 245, 245, 1);
+  static final blackHint = Color.fromRGBO(61, 61, 61, 1);
+  static final shadowColor = Color.fromRGBO(143, 143, 143, 1);
+  static final iconOutlineColor = Color.fromRGBO(218, 220, 224, 1);
+}
