+++ lib/home/LabelWIthActionButton.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 
 // ignore: must_be_immutable
 class LabelWithActionButton extends StatelessWidget {
