+++ lib/home/SearchBar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 import 'package:grocery/SearchPage/SearchPage.dart';
 
 // ignore: must_be_immutable
