+++ lib/product/WideProductCard.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 import 'package:grocery/custom_widget/Button.dart';
 import 'package:http/http.dart' as http;
 import 'ProductCard.dart';
