+++ lib/products/ProductGroupGridItems.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:grocery/HttpRequestService.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 import 'package:grocery/home/BottomNavBar.dart';
 import 'package:grocery/products/LabelWithActionButton.dart';
 import 'package:grocery/product/ProductCard.dart';
