+++ lib/profile/Register.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:grocery/constrant.dart';
+import 'package:grocery/constant.dart';
 
 class Register extends StatefulWidget {
   static String routeName = '/register';
